requirements.txt
email-validator>=2.2,<3.0
Flask>=3.1,<4.0
Flask-Limiter>=3.7,<4.0
Flask-Login>=0.6.3,<1.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-WTF>=1.2,<2.0
groq>=0.18,<1.0
gunicorn>=23.0,<24.0
python-dotenv>=1.0,<2.0
Werkzeug>=3.1,<4.0
